FixedComp : Re-Alignment of axis
    In some models it is useful to realign a FixedComp object to a
    particular axis. E.g. a chopper pit / shielding is likely to be
    aligned with Z == (0,0,1). But the object needs to be build relative
    to an object on a slope (e.g. a beamline/guide etc). To facilitate
    this FixedComp has three new methods that are accessed via two
    public interfaces.

    setAxisControl(long int,Vec3D)

    This takes an axis number (0 for no op) 1-3 for X,Y,Z that
    will be aligned with  the given vec3D. The initial index can be
    negative/positive. Rotation is carried out on all principle axes via a
    rotation about the normal vector to the plane that contains both the
    old axis and the new axis.

    positive: the reorientation is applied
    before any xstep/rotation shift to be applied via variables.

    negitive: the reorientation is applied
    after any xstep/rotation shift to be applied via variables.

    The system is propogated down via FixedGroup / FixedOffset /
     FixedGroupOffset.

    Specific reorientations can be accessed via
    reOrientate() and reOrientate(size_t,Vec3D).

    reOrientate  uses the axis number (0-2) and the normal vector given.
    If it is not given then the existing values set via setAxisControl
    are used.

    Test case used and examples are in FREIA beamline.  [FREIA.cxx]

+++ System/attachCompInc/FixedGroup.h
@@ -39,7 +39,7 @@ class FixedGroup : public FixedComp
   /// Fixed Storage
   typedef std::shared_ptr<attachSystem::FixedComp> CompTYPE;
   /// Map storage
-  typedef std::map<std::string,CompTYPE > FTYPE;
+  typedef std::map<std::string,CompTYPE> FTYPE;
   /// Named Container
   FTYPE FMap;
 
@@ -68,7 +68,7 @@ class FixedGroup : public FixedComp
   virtual const FixedComp& getKey(const std::string&) const;
 
   
-  void setAxisControl(const size_t,const Geometry::Vec3D&);
+  void setAxisControl(const long int,const Geometry::Vec3D&);
 
 };
